+++ WorldBuilder.Shared/Modules/Landscape/Models/SceneRaycastHit.cs
@@ -25,7 +25,7 @@ namespace WorldBuilder.Shared.Modules.Landscape.Models {
         public uint LandblockId { get; set; }
         
         /// <summary>The instance ID of the hit object.</summary>
-        public uint InstanceId { get; set; }
+        public ulong InstanceId { get; set; }
         
         /// <summary>The object ID (from dats) of the hit object.</summary>
         public uint ObjectId { get; set; }
